requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apturl==0.5.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.1.0
Click==7.0
cmake==3.26.3
colorama==0.4.3
command-not-found==0.3
contourpy==1.0.7
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
datasets==2.11.0
dbus-python==1.2.16
defer==1.0.6
dill==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.2
httplib2==0.14.0
huggingface-hub==0.13.4
idna==2.8
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lit==16.0.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.2
matplotlib==3.7.1
monotonic==1.5
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
netifaces==0.10.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
plotly==5.14.1
protobuf==3.20.3
pyarrow==11.0.0
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
python-apt==2.0.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2023.3
pyxdg==0.26
PyYAML==5.3.1
regex==2023.3.23
reportlab==3.5.34
requests==2.28.2
requests-unixsocket==0.2.0
responses==0.18.0
scikit-learn==1.2.2
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
shapely==2.0.1
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
sympy==1.11.1
systemd-python==234
tenacity==8.2.2
tensorboardX==2.6
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing-extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.15
usb-creator==0.3.7
wadllib==1.3.3
xkit==0.0.0
xxhash==3.2.0
yarl==1.9.1
zipp==3.15.0
